# docs/swagger/swagger.yaml
basePath: /core/v1
definitions:
...
  lakukan-be_internal_modules_crm_company_settings_dto.UpdateCompanySettingsRequest:
    properties:
      ai_confidence_threshold:
        maximum: 1
        minimum: 0
        type: number
      ai_enabled:
        description: AI Settings
        type: boolean
      ai_training_mode:
        type: boolean
      auto_reply_business_hours_end:
        description: 'Format: "HH:MM:SS"'
        type: string
      auto_reply_business_hours_start:
        description: 'Format: "HH:MM:SS"'
        type: string
      auto_reply_enabled:
        description: Auto-Reply Settings
        type: boolean
      auto_reply_global_delay_seconds:
        minimum: 0
        type: integer
      auto_reply_working_days:
        items:
          type: integer
        type: array
      chat_assignment_strategy:
        enum:
        - round_robin
        - least_busy
        - manual
        type: string
      chat_auto_assign_enabled:
        description: Chat Settings
        type: boolean
      chat_idle_timeout_minutes:
        minimum: 1
        type: integer
      lead_auto_categorization_enabled:
        description: Lead Settings
        type: boolean
      lead_cold_criteria: {}
      lead_hot_criteria: {}
      lead_warm_criteria: {}
      notification_channels: {}
      notifications_enabled:
        description: Notification Settings
        type: boolean
      timezone:
        description: Timezone
        type: string
    type: object
  lakukan-be_internal_modules_crm_lead_sources_dto.CreateLeadSourceRequest:
    properties:
      code:
        maxLength: 50
        minLength: 2
        type: string
      color:
        description: '#FFFFFF format'
        type: string
      description:
        maxLength: 500
        type: string
      icon:
        maxLength: 50
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 3
        type: string
      sort_order:
        minimum: 0
        type: integer
    required:
    - code
    - name
    type: object
  lakukan-be_internal_modules_crm_lead_sources_dto.UpdateLeadSourceRequest:
    properties:
      code:
        maxLength: 50
        minLength: 2
        type: string
      color:
        type: string
      description:
        maxLength: 500
        type: string
      icon:
        maxLength: 50
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 3
        type: string
      sort_order:
        minimum: 0
        type: integer
    type: object
...
  lakukan-be_internal_shared_response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/lakukan-be_internal_shared_response.Pagination'
    type: object
  lakukan-be_internal_shared_response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  lakukan-be_internal_shared_response.Response:
    properties:
      data: {}
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        type: string
      meta:
        $ref: '#/definitions/lakukan-be_internal_shared_response.Meta'
    type: object
host: localhost:8080
info:
  contact:
    email: support@lakukan.com
    name: API Support
    url: http://www.lakukan.com/support
  description: Backend API untuk Lakukan - Multi-tenant Business Management Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Lakukan API
  version: "1.0"
paths:
...
  /crm/lead-sources:
    get:
      consumes:
      - application/json
      description: Get list of lead sources with pagination and filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by name, code, or description
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Lead sources retrieved successfully
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get all lead sources
      tags:
      - CRM - Lead Sources
    post:
      consumes:
      - application/json
      description: Create a new lead source
      parameters:
      - description: Lead source data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_crm_lead_sources_dto.CreateLeadSourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Lead source created successfully
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "409":
          description: Lead source code already exists
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Create new lead source
      tags:
      - CRM - Lead Sources
  /crm/lead-sources/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a lead source
      parameters:
      - description: Lead Source ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lead source deleted successfully
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Lead source not found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Delete lead source
      tags:
      - CRM - Lead Sources
    get:
      consumes:
      - application/json
      description: Get lead source details by ID
      parameters:
      - description: Lead Source ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lead source retrieved successfully
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Lead source not found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get lead source by ID
      tags:
      - CRM - Lead Sources
    put:
      consumes:
      - application/json
      description: Update an existing lead source
      parameters:
      - description: Lead Source ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated lead source data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_crm_lead_sources_dto.UpdateLeadSourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Lead source updated successfully
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Lead source not found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "409":
          description: Lead source code already exists
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Update lead source
      tags:
      - CRM - Lead Sources
  /crm/lead-sources/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restore a soft-deleted lead source
      parameters:
      - description: Lead Source ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lead source restored successfully
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Lead source not found or not deleted
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Restore lead source
      tags:
      - CRM - Lead Sources
...
  /crm/v1/settings:
    get:
      consumes:
      - application/json
      description: Get CRM settings for the current company (creates default if not
        exists)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get company CRM settings
      tags:
      - CRM - Company Settings
    put:
      consumes:
      - application/json
      description: Update CRM settings for the current company
      parameters:
      - description: Settings data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_crm_company_settings_dto.UpdateCompanySettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Update company CRM settings
      tags:
      - CRM - Company Settings
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"